tctl resource: convert static host user handler
