+++ tool/tctl/common/collection.go
@@ -22,7 +22,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"slices"
 	"sort"
 	"strconv"
 	"strings"
@@ -37,7 +36,6 @@ import (
 	healthcheckconfigv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/healthcheckconfig/v1"
 	loginrulepb "github.com/gravitational/teleport/api/gen/proto/go/teleport/loginrule/v1"
 	scopedaccessv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/scopes/access/v1"
-	userprovisioningpb "github.com/gravitational/teleport/api/gen/proto/go/teleport/userprovisioning/v2"
 	usertasksv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/usertasks/v1"
 	"github.com/gravitational/teleport/api/gen/proto/go/teleport/vnet/v1"
 	"github.com/gravitational/teleport/api/types"
@@ -1026,64 +1024,6 @@ func (c *pluginCollection) WriteText(w io.Writer, verbose bool) error {
 	}
 	_, err := t.AsBuffer().WriteTo(w)
 	return trace.Wrap(err)
-}
-
-type staticHostUserCollection struct {
-	items []*userprovisioningpb.StaticHostUser
-}
-
-func (c *staticHostUserCollection) Resources() []types.Resource {
-	r := make([]types.Resource, 0, len(c.items))
-	for _, resource := range c.items {
-		r = append(r, types.Resource153ToLegacy(resource))
-	}
-	return r
-}
-
-func (c *staticHostUserCollection) WriteText(w io.Writer, verbose bool) error {
-	var rows [][]string
-	for _, item := range c.items {
-
-		for _, matcher := range item.Spec.Matchers {
-			labelMap := label.ToMap(matcher.NodeLabels)
-			labelStringMap := make(map[string]string, len(labelMap))
-			for k, vals := range labelMap {
-				labelStringMap[k] = fmt.Sprintf("[%s]", printSortedStringSlice(vals))
-			}
-			var uid string
-			if matcher.Uid != 0 {
-				uid = strconv.Itoa(int(matcher.Uid))
-			}
-			var gid string
-			if matcher.Gid != 0 {
-				gid = strconv.Itoa(int(matcher.Gid))
-			}
-			rows = append(rows, []string{
-				item.GetMetadata().Name,
-				common.FormatLabels(labelStringMap, verbose),
-				matcher.NodeLabelsExpression,
-				printSortedStringSlice(matcher.Groups),
-				uid,
-				gid,
-			})
-		}
-	}
-	headers := []string{"Login", "Node Labels", "Node Expression", "Groups", "Uid", "Gid"}
-	var t asciitable.Table
-	if verbose {
-		t = asciitable.MakeTable(headers, rows...)
-	} else {
-		t = asciitable.MakeTableWithTruncatedColumn(headers, rows, "Node Expression")
-	}
-	t.SortRowsBy([]int{0}, true)
-	_, err := t.AsBuffer().WriteTo(w)
-	return trace.Wrap(err)
-}
-
-func printSortedStringSlice(s []string) string {
-	s = slices.Clone(s)
-	slices.Sort(s)
-	return strings.Join(s, ",")
 }
 
 type userTaskCollection struct {
